+++ tests/projects/common/account/views.py
@@ -1,0 +1,8 @@
+from django.http import HttpResponse
+
+from allauth.account.decorators import verified_email_required
+
+
+@verified_email_required
+def check_verified_email(request):
+    return HttpResponse("VERIFIED")